API

Library

Module

AugmentedGaussianProcesses.AugmentedGaussianProcesses — Module.

General Framework for the data augmented Gaussian Processes

source

Model Types

GP(X,y,kernel)
VGP(X,y,kernel,likelihood,inference)
SVGP(X,y,kernel,likelihood,inference,num_ind_points)

Likelihood Types

GaussianLikelihood()
AugmentedStudentTLikelihood(ν)
BayesianSVM()
AugmentedLogisticLikelihood()
AugmentedLogisticSoftMaxLikelihood()

Inference Types

AnalyticInference()
StochasticAnalyticInference(batchsize)
NumericalInference(optimizer)
StochasticNumericalInference(batchsize)
GibbsSampling()

Functions and methods

train!(model)
predict_f(model,X_test)
predict_y(model,X_test)
proba_y(model,X_test)

Kernels

RBFKernel()
MaternKernel)_

Kernel functions

AugmentedGaussianProcesses.KernelModule.kernelmatrix — Function.

Create the covariance matrix between the matrix X1 and X2 with the covariance function kernel

source

Compute the covariance matrix of the matrix X, optionally only compute the diagonal terms

source

AugmentedGaussianProcesses.KernelModule.kernelmatrix! — Function.

Compute the covariance matrix between the matrix X1 and X2 with the covariance function kernel in preallocated matrix K

source

Compute the covariance matrix of the matrix X in preallocated matrix K, optionally only compute the diagonal terms

source

AugmentedGaussianProcesses.KernelModule.getvariance — Function.

Return the variance of the kernel

source

AugmentedGaussianProcesses.KernelModule.getlengthscales — Function.

Return the lengthscale of the IsoKernel

source

Return the lengthscales of the ARD Kernel

source

Index

AugmentedGaussianProcesses.KernelModule.getlengthscales

AugmentedGaussianProcesses.KernelModule.getvariance

AugmentedGaussianProcesses.KernelModule.kernelmatrix

AugmentedGaussianProcesses.KernelModule.kernelmatrix!