API

API Library

API Library

Module

Model Types

Likelihood Types

Inference Types

Functions and methods

Kernels

Kernel functions

Index

Module

AugmentedGaussianProcesses.AugmentedGaussianProcesses — Module.

General Framework for the data augmented Gaussian Processes

source

Model Types

AugmentedGaussianProcesses.GP — Type.

Class for variational Gaussian Processes models (non-sparse)

source

AugmentedGaussianProcesses.VGP — Type.

Class for variational Gaussian Processes models (non-sparse)

source

AugmentedGaussianProcesses.SVGP — Type.

Class for sparse variational Gaussian Processes

source

Likelihood Types

AugmentedGaussianProcesses.GaussianLikelihood — Type.

Gaussian likelihood : $p(y|f) = \mathcal{N}(y|f,\epsilon)$

source

AugmentedGaussianProcesses.StudentTLikelihood — Type.

Student-T likelihood

Student-t likelihood for regression: $\frac{\Gamma((\nu+1)/2)}{\sqrt{\nu\pi}\Gamma(\nu/2)}\left(1+t^2/\nu\right)^{(-(\nu+1)/2)}$ see wiki page

For the analytical solution, it is augmented via:

\[#TODO\]

See paper Robust Gaussian Process Regression with a Student-t Likelihood

source

AugmentedGaussianProcesses.BayesianSVM — Type.

The Bayesian SVM is a Bayesian interpretation of the classical SVM. By using an augmentation (Laplace) one gets a conditionally conjugate likelihood (see paper)

source

AugmentedGaussianProcesses.LogisticLikelihood — Type.

Logistic Likelihood

Bernoulli likelihood with a logistic link for the Bernoulli likelihood $p(y|f) = \sigma(yf) = \frac{1}{1+\exp(-yf)}$, (for more info see : wiki page)

---

For the analytic versionm the likelihood is augmented to give a conditionally conjugate likelihood :
```math
p(y|f,\omega) = \exp\left(\frac{1}{2}\left(yf - (yf)^2 \omega\right)\right)
```
where ``\omega \sim \text{PG}(\omega\mid 1, 0)``, and PG is the Polya-Gamma distribution
See paper : [Efficient Gaussian Process Classification Using Polya-Gamma Data Augmentation](https://arxiv.org/abs/1802.06383)

source

AugmentedGaussianProcesses.LogisticSoftMaxLikelihood — Type.

The Logistic-Softmax likelihood

The multiclass likelihood with a logistic-softmax mapping: : $p(y=i|\{f_k\}) = \sigma(f_i)/ \sum_k \sigma(f_k)$ where σ is the logistic function has the same properties as softmax.

For the analytical version, the likelihood is augmented multiple times to obtain :

\[#TODO\]

Paper with details under submission

source

Inference Types

AugmentedGaussianProcesses.AnalyticVI — Type.

Solve conjugate or conditionally conjugate likelihoods (especially valid for augmented likelihoods)

source

AugmentedGaussianProcesses.AnalyticSVI — Function.

AnalyticSVI(nMinibatch::Integer;ϵ::T=1e-5,optimizer::Optimizer=ALRSVI())

Return an AnalyticVI{T} object with stochastic updates, corresponding to Stochastic Variational Inference with analytical updates.

Positional argument

- `nMinibatch::Integer` : Number of samples per mini-batches

Keywords arguments

- `ϵ::T` : convergence criteria, which can be user defined
- `optimizer::Optimizer` : Optimizer used for the variational updates. Should be an Optimizer object from the [GradDescent.jl]() package. Default is `ALRSVI()` (Adaptive Learning Rate for Stochastic Variational Inference)

source

AugmentedGaussianProcesses.NumericalVI — Type.

Solve any non-conjugate likelihood using Variational Inference by making a numerical approximation (quadrature or MC integration) of the expected log-likelihood ad its gradients

source

AugmentedGaussianProcesses.NumericalSVI — Function.

NumericalSVI(integration_technique::Symbol=:quad;ϵ::T=1e-5,nMC::Integer=1000,nGaussHermite::Integer=20,optimizer::Optimizer=Adam(α=0.1))

General constructor for Stochastic Variational Inference via numerical approximation.

Argument

-`nMinibatch::Integer` : Number of samples per mini-batches
-`integration_technique::Symbol` : Method of approximation can be `:quad` for quadrature see [QuadratureVI](@ref) or `:mcmc` for MCMC integration see [MCMCIntegrationVI](@ref)

Keyword arguments

- `ϵ::T` : convergence criteria, which can be user defined
- `nMC::Int` : Number of samples per data point for the integral evaluation (for the MCMCIntegrationVI)
- `nGaussHermite::Int` : Number of points for the integral estimation (for the QuadratureVI)
- `optimizer::Optimizer` : Optimizer used for the variational updates. Should be an Optimizer object from the [GradDescent.jl]() package. Default is `Adam()`

source

AugmentedGaussianProcesses.QuadratureVI — Type.

QuadratureVI(integration_technique::Symbol=:quad;ϵ::T=1e-5,nGaussHermite::Integer=20,optimizer::Optimizer=Adam(α=0.1))

Constructor for Variational Inference via quadrature approximation.

Keyword arguments

- `ϵ::T` : convergence criteria, which can be user defined
- `nGaussHermite::Int` : Number of points for the integral estimation (for the QuadratureVI)
- `optimizer::Optimizer` : Optimizer used for the variational updates. Should be an Optimizer object from the [GradDescent.jl]() package. Default is `Adam()`

source

AugmentedGaussianProcesses.QuadratureSVI — Function.

QuadratureSVI(;ϵ::T=1e-5,nMC::Integer=1000,optimizer::Optimizer=Adam(α=0.1))

Constructor for Stochastic Variational Inference via quadrature approximation.

Argument

-`nMinibatch::Integer` : Number of samples per mini-batches

Keyword arguments

- `ϵ::T` : convergence criteria, which can be user defined
- `nGaussHermite::Int` : Number of points for the integral estimation (for the QuadratureVI)
- `optimizer::Optimizer` : Optimizer used for the variational updates. Should be an Optimizer object from the [GradDescent.jl]() package. Default is `Adam()`

source

AugmentedGaussianProcesses.MCMCIntegrationVI — Type.

MCMCIntegrationVI(integration_technique::Symbol=:quad;ϵ::T=1e-5,nMC::Integer=1000,optimizer::Optimizer=Adam(α=0.1))

Constructor for Variational Inference via MCMC Integration approximation.

Keyword arguments

- `ϵ::T` : convergence criteria, which can be user defined
- `nMC::Int` : Number of samples per data point for the integral evaluation
- `optimizer::Optimizer` : Optimizer used for the variational updates. Should be an Optimizer object from the [GradDescent.jl]() package. Default is `Adam()`

source

AugmentedGaussianProcesses.MCMCIntegrationSVI — Function.

MCMCIntegrationSVI(;ϵ::T=1e-5,nMC::Integer=1000,optimizer::Optimizer=Adam(α=0.1))

Constructor for Stochastic Variational Inference via MCMC integration approximation.

Argument

-`nMinibatch::Integer` : Number of samples per mini-batches

Keyword arguments

- `ϵ::T` : convergence criteria, which can be user defined
- `nMC::Int` : Number of samples per data point for the integral evaluation
- `optimizer::Optimizer` : Optimizer used for the variational updates. Should be an Optimizer object from the [GradDescent.jl]() package. Default is `Adam()`

source

AugmentedGaussianProcesses.GibbsSampling — Type.

GibbsSampling(;ϵ::T=1e-5,nBurnin::Int=100,samplefrequency::Int=10)

Return a GibbsSampling{T} object to sample from the exact posterior distribution.

Keywords arguments

- `ϵ::T` : convergence criteria, which can be user defined
- `nBurnin::Int` : Number of samples discarded before starting to save samples
- `samplefrequency::Int` : Frequency of sampling

source

Functions and methods

AugmentedGaussianProcesses.train! — Function.

train!(model::AbstractGP;iterations::Integer=100,callback=0,conv_function=0)

Function to train the given GP model.

Keyword Arguments

there are options to change the number of max iterations,

iterations::Int : Number of iterations (not necessarily epochs!)for training

callback::Function : Callback function called at every iteration. Should be of type function(model,iter) ... end

conv_function::Function : Convergence function to be called every iteration, should return a scalar and take the same arguments as callback

source

AugmentedGaussianProcesses.predict_f — Function.

Compute the mean of the predicted latent distribution of f on X_test for the variational GP model

Return also the variance if covf=true and the full covariance if fullcov=true

source

Compute the mean of the predicted latent distribution of f on X_test for a sparse GP model Return also the variance if covf=true and the full covariance if fullcov=true

source

AugmentedGaussianProcesses.predict_y — Function.

predict_y(model::AbstractGP{<:RegressionLikelihood},X_test::AbstractMatrix)

Return the predictive mean of X_test

source

predict_y(model::AbstractGP{<:ClassificationLikelihood},X_test::AbstractMatrix)

Return the predicted most probable sign of X_test

source

predict_y(model::AbstractGP{<:MultiClassLikelihood},X_test::AbstractMatrix)

Return the predicted most probable class of X_test

source

AugmentedGaussianProcesses.proba_y — Function.

proba_y(model::AbstractGP,X_test::AbstractMatrix)

Return the probability distribution p(ytest|model,Xtest) :

- Tuple of vectors of mean and variance for regression
- Vector of probabilities of y_test = 1 for binary classification
- Dataframe with columns and probability per class for multi-class classification

source

Kernels

AugmentedGaussianProcesses.KernelModule.RBFKernel — Type.

Radial Basis Function Kernel also called RBF or SE(Squared Exponential)

source

AugmentedGaussianProcesses.KernelModule.MaternKernel — Type.

Matern Kernel

source

Kernel functions

AugmentedGaussianProcesses.KernelModule.kernelmatrix — Function.

Create the covariance matrix between the matrix X1 and X2 with the covariance function kernel

source

Compute the covariance matrix of the matrix X, optionally only compute the diagonal terms

source

AugmentedGaussianProcesses.KernelModule.kernelmatrix! — Function.

Compute the covariance matrix between the matrix X1 and X2 with the covariance function kernel in preallocated matrix K

source

Compute the covariance matrix of the matrix X in preallocated matrix K, optionally only compute the diagonal terms

source

AugmentedGaussianProcesses.KernelModule.getvariance — Function.

Return the variance of the kernel

source

AugmentedGaussianProcesses.KernelModule.getlengthscales — Function.

Return the lengthscale of the IsoKernel

source

Return the lengthscales of the ARD Kernel

source

Index

AugmentedGaussianProcesses.AnalyticVI

AugmentedGaussianProcesses.BayesianSVM

AugmentedGaussianProcesses.GP

AugmentedGaussianProcesses.GaussianLikelihood

AugmentedGaussianProcesses.GibbsSampling

AugmentedGaussianProcesses.KernelModule.MaternKernel

AugmentedGaussianProcesses.KernelModule.RBFKernel

AugmentedGaussianProcesses.LogisticLikelihood

AugmentedGaussianProcesses.LogisticSoftMaxLikelihood

AugmentedGaussianProcesses.MCMCIntegrationVI

AugmentedGaussianProcesses.NumericalVI

AugmentedGaussianProcesses.QuadratureVI

AugmentedGaussianProcesses.SVGP

AugmentedGaussianProcesses.StudentTLikelihood

AugmentedGaussianProcesses.VGP

AugmentedGaussianProcesses.AnalyticSVI

AugmentedGaussianProcesses.KernelModule.getlengthscales

AugmentedGaussianProcesses.KernelModule.getvariance

AugmentedGaussianProcesses.KernelModule.kernelmatrix

AugmentedGaussianProcesses.KernelModule.kernelmatrix!

AugmentedGaussianProcesses.MCMCIntegrationSVI

AugmentedGaussianProcesses.NumericalSVI

AugmentedGaussianProcesses.QuadratureSVI

AugmentedGaussianProcesses.predict_f

AugmentedGaussianProcesses.predict_y

AugmentedGaussianProcesses.proba_y

AugmentedGaussianProcesses.train!